rev35
Because aPortGenerationWithWorker works in Firefox 46+, I added shared
scope method. I created discussion topic on it here -

http://discourse.mozilla-community.org/t/tap-into-content-scope-or-transfer-ports-from-webworker/10858

In Firefox < 46, the ports I obtain from worker, then send to content,
are not usable. Window/`content` silently fails to do postMessage of
this[messager_method]('contentComm_handshake_finalized'); - it just
wouldnt do it. if i make it from shared scope though it is able to send
down it. (framescript is able to send postMessage down it, and it
`content` does receive the message, its just that `content` cant send
out)